Cleaned up some of the cruft.  Changed utility central so it gets an event structure as the parameter list was getting out of hand.
Right now you can select in the palette around where the shapes
should be drawn (get get some selectedShape set) then you can
clickAndDrag the mouse in the paint area to put white rectangles
on the canvas for now.  Lots more to do but this is starting
to shape up.